+++ src/main/java/com/paymenttransaction/payment_transaction_manager/transactions/application/ports/out/card/CardPort.java
@@ -1,5 +1,6 @@
 package com.paymenttransaction.payment_transaction_manager.transactions.application.ports.out.card;
 
+import com.paymenttransaction.payment_transaction_manager.transactions.domain.enums.TransactionStatus;
 import com.paymenttransaction.payment_transaction_manager.transactions.domain.models.Card;
 
 import java.util.List;
@@ -9,4 +10,5 @@ public interface CardPort {
     Card createCard(Card card);
     Optional<Card> getCardById(Long id);
     Optional<List<Card>> getCards(Long userId);
+    TransactionStatus getTransactionStatusById(Long id);
 }