+++ src/main/java/com/paymenttransaction/payment_transaction_manager/transactions/infrastructure/adapters/TransactionAdapter.java
@@ -2,10 +2,12 @@ package com.paymenttransaction.payment_transaction_manager.transactions.infrastr
 
 
 import com.paymenttransaction.payment_transaction_manager.transactions.application.ports.out.TransactionPort;
+import com.paymenttransaction.payment_transaction_manager.transactions.domain.enums.TransactionStatus;
 import com.paymenttransaction.payment_transaction_manager.transactions.domain.models.Transaction;
 import com.paymenttransaction.payment_transaction_manager.transactions.infrastructure.entities.TransactionEntity;
 import com.paymenttransaction.payment_transaction_manager.transactions.infrastructure.mappers.TransactionEntityMapper;
 import com.paymenttransaction.payment_transaction_manager.transactions.infrastructure.repositories.TransactionRepository;
+import org.springframework.cache.annotation.Cacheable;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
@@ -36,8 +38,13 @@ public class TransactionAdapter implements TransactionPort {
     }
 
     @Override
+    @Cacheable(value = "transaction", key = "#id")
     public Optional<Transaction> getTransactionById(Long id) {
         return repository.findById(id).map(mapper::toModel);
+    }
+
+    public TransactionStatus getTransactionStatusById(Long id) {
+        return repository.findById(id).map(TransactionEntity::getStatus).orElse(null);
     }
 }