+++ CHANGELOG.md
@@ -6,6 +6,21 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [3.1.0] - 2025-01-28
+
+### Added
+
+- Cache logic.
+- Find Transaction status logic.
+
+### Changed
+
+- Classes and tests.
+
+### Fixed
+
+- Fixed bugs on cards, bank and p2p transactions.
+
 ## [3.0.0] - 2025-01-28
 
 ### Changed